requirements.txt
pillow
pytesseract
PyMuPDF
openai
bottle